requirements.txt
# Django
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
django-environ==0.9.0

# Database
psycopg2-binary==2.9.7
dj-database-url==2.1.0

# File handling
Pillow==10.0.1
whitenoise==6.6.0

# Production
gunicorn==21.2.0

# Environment
python-decouple==3.8